Mvc controller and view for cards have been added
