+++ LanguageCards.WebApp/Views/Cards/Index.cshtml
@@ -1,0 +1,19 @@
+﻿
+@{
+    Layout = null;
+}
+
+<!DOCTYPE html>
+<html>
+<head>
+    <meta name="viewport" content="width=device-width" />
+    <meta charset="utf-8" />
+    <title>Index</title>
+    <link rel="stylesheet" href="~/dist/vendor.css" />
+</head>
+<body>
+    <h2>Hello!</h2>
+    @*<div id="app">Loading...</div>*@
+    <counter>Loading...</counter>
+</body>
+</html>
